Initial BGR2RGB Swizzle testing

+++ src/main/java/net/vulkanmod/vulkan/memory/StagingBuffer.java
@@ -29,11 +29,11 @@ public class StagingBuffer extends Buffer{
     public void copyBuffer(int size, ByteBuffer byteBuffer) {
 
         if(size > this.bufferSize - this.usedBytes) {
-            resizeBuffer((int)(this.bufferSize + size) * 2);
+            resizeBuffer((this.bufferSize + size) * 2);
         }
 
         MemoryManager.getInstance().Copy(this.data,
-                (data) -> VUtil.memcpy(data.getByteBuffer(0, (int) this.bufferSize), byteBuffer, this.usedBytes)
+                (data) -> VUtil.memcpy(data.getByteBuffer(0, this.bufferSize), byteBuffer, this.usedBytes)
         );
 
         offset = usedBytes;
@@ -54,7 +54,7 @@ public class StagingBuffer extends Buffer{
         usedBytes = 0;
     }
 
-    public long getOffset() {
+    public int getOffset() {
         return offset;
     }
